% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_neighbors.R
\name{fast_make_all_neighbors}
\alias{fast_make_all_neighbors}
\title{Make table with all cell neighbors within a radius.}
\usage{
fast_make_all_neighbors(
  dt,
  cellid_col = "cell_ID",
  sdimx_col = "sdimx",
  sdimy_col = "sdimy",
  radius = 0.05
)
}
\arguments{
\item{dt}{data.table of metadata containing (at least) columns for cell ids,
x-coordinates, y-coordinates.}

\item{cellid_col}{name of column with cellid, i.e., "cell_ID".}

\item{sdimx_col}{name of column with x-coordinate for cell, i.e., "sdimx".}

\item{sdimy_col}{name of column with x-coordinate for cell, i.e., "sdimy".}

\item{radius}{radius within which to identify all other neighboring cells, i.e, 0.05.}
}
\description{
Make table with all cell neighbors within a radius.
Passed to downstream functions
such as \code{smiDE::measure_neighbor_expression_by_celltype}.
}
\examples{
library(Giotto)
library(data.table)
datadir<-system.file("extdata", package="smiDE")
gem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
metainfo <- data.table::copy(gem@cell_metadata$rna)
metainfo <- merge(metainfo, gem@spatial_locs$raw, by="cell_ID")

tiss <- split(metainfo, by="tissue")

cell_adjacency_dt <- 
  rbindlist(
    lapply(1:length(tiss), function(x){
      nb <- fast_make_all_neighbors(tiss[[x]][,.(cell_ID, sdimx, sdimy)]
                                    ,cellid_col = "cell_ID"
                                    ,sdimx_col ="sdimx"
                                    ,sdimy_col ="sdimy"
                                    ,radius = 0.05) 
      return(nb) 
    })
  )

}
