% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mod_emmeans.R
\name{make_cov_reduce_list}
\alias{make_cov_reduce_list}
\title{Function for creating a input list for emmeans cov.reduce argument}
\usage{
make_cov_reduce_list(fixedterms, dat)
}
\description{
Function for creating a input list for emmeans cov.reduce argument
}
