% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_de.R
\name{pre_de}
\alias{pre_de}
\title{Calculate cell adjacencies and neighboring cell expression of genes before DE analysis.}
\usage{
pre_de(
  counts,
  normalized_data = NULL,
  metadata,
  ref_celltype,
  cell_type_metadata_colname,
  weight_colname = NULL,
  cellid_colname = "cell_ID",
  sdimx_colname = "sdimx",
  sdimy_colname = "sdimy",
  split_neighbors_by_colname = "tissue",
  mm_radius = 0.05,
  verbose = TRUE,
  contamination = "sum",
  nb_celltypes_to_individually_calc = NULL
)
}
\arguments{
\item{counts}{matrix of counts}

\item{normalized_data}{optional, matrix of normalized data.
If not provided, totalCounts normalization is applied to counts.}

\item{metadata}{cell-level metadata corresponding to the cells in counts, with one row per cell.
Should contain the columns of cell_type_metadata_colname,
cellid_colname, split_neighbors_by_colname, sdimx_colname, and sdimy_colname.}

\item{ref_celltype}{Reference category (commonly a cell type), used to calculate neighbor expression with respect to.
Can be vector of one or more levels of cell_type_metadata_colname.
Can use keyword "all" for all levels of cell_type_metadata_colname.}

\item{cell_type_metadata_colname}{column (commonly cell type column) corresponding to categories by which cell expression is
aggregated by to calculate expression among neighboring cells by cell_type_metadata_colname type.}

\item{weight_colname}{(optional).  If NULL, neighbor expression is unweighted.
If "weight", the column corresponding to "weight" (1/distance) by default is used to weight calculated expression of neighbor cells.}

\item{cellid_colname}{column name corresponding to cell id.}

\item{sdimx_colname}{column name corresponding to 'x' axis spatial coordinate of cell.}

\item{sdimy_colname}{column name corresponding to 'y' axis spatial coordinate of cell.}

\item{split_neighbors_by_colname}{(optional) If specified, identification of neighboring cells is split by this
column in the meta data.  For example, to avoid identifying cells as neighbors if they were to have close x/y coordinates, but were on different tissue.}

\item{mm_radius}{maximum euclidean distance in units of sdimx, sdimy by which two cells will be identified as neighbors.}

\item{verbose}{print some extra messages during calculation.}

\item{contamination}{"sum" (default) calculates a total amount of target expressed in neighboring cells.  Anything other values passed will calculate an average.}

\item{nb_celltypes_to_individually_calc}{vector of cell types for which to individually calculate neighbor expression.
default is NULL, which will make a separate list in neighbor_expr_by_ct output with matrix of expression by cell type}
}
\value{
list object with data.table of cell_adjacencies (one row per pair of adjacent cells),
a list (nblist) with sparse matrices corresponding to neighbor cell expression passed to downstream DE in smi_de or run_de functions.
}
\description{
Calculate cell adjacencies and neighboring cell expression of genes before DE analysis.
}
\examples{
library(Giotto)
library(data.table)
datadir<-system.file("extdata", package="smiDE")
gem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
metainfo <- data.table::copy(gem@cell_metadata$rna)
metainfo <- merge(metainfo, gem@spatial_locs$raw, by="cell_ID")
 
pre_de_obj <- 
pre_de(counts = gem@expression[["rna"]][["raw"]]
       ,normalized_data = gem@expression[["rna"]][["normalized"]]
       ,metadata = metainfo
       ,cell_type_metadata_colname = "cell_type"
       ,split_neighbors_by_colname = "tissue"
       ,mm_radius = 0.05
       ,ref_celltype = "fibroblast"
       ,sdimx_colname = "sdimx"
       ,sdimy_colname = "sdimy"
       ,weight_colname = "weight"
       ,contamination = "sum"
       ,verbose=TRUE
 )
names(pre_de_obj)

# Example for measuring neighbor expression with 'all' levels of
# ref_celltype = 'all'
pre_de_obj <- 
pre_de(counts = gem@expression[["rna"]][["raw"]]
       ,normalized_data = gem@expression[["rna"]][["normalized"]]
       ,metadata = metainfo
       ,cell_type_metadata_colname = "cell_type"
       ,split_neighbors_by_colname = "tissue"
       ,mm_radius = 0.05
       ,ref_celltype = c("all")
       ,sdimx_colname = "sdimx"
       ,sdimy_colname = "sdimy"
       ,weight_colname = "weight"
       ,contamination = "sum"
       ,verbose=TRUE
)
names(pre_de_obj)
dim(pre_de_obj$nblist$neighbor_expr_byct$otherct)  
metainfo[, .N]


# Example for measuring neighbor expression with a vector of multiple cell types
pre_de_obj <- 
pre_de(counts = gem@expression[["rna"]][["raw"]]
       ,normalized_data = gem@expression[["rna"]][["normalized"]]
       ,metadata = metainfo
       ,cell_type_metadata_colname = "cell_type"
       ,split_neighbors_by_colname = "tissue"
       ,mm_radius = 0.05
       ,ref_celltype = c("macrophage", "fibroblast")
       ,sdimx_colname = "sdimx"
       ,sdimy_colname = "sdimy"
       ,weight_colname = "weight"
       ,contamination = "sum"
       ,verbose=TRUE
)
names(pre_de_obj)
dim(pre_de_obj$nblist$neighbor_expr_byct$otherct)  
metainfo[cell_type \%in\% pre_de_obj$nblist$ref_celltype, .N]
 
}
