% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mod_emmeans.R
\name{prepare_mod_emmeans}
\alias{prepare_mod_emmeans}
\title{Function for converting nebula model to emmeans object}
\usage{
prepare_mod_emmeans(mod, term, fitmethod, fixed_formula, dat, cov_reduce_list)
}
\description{
Function for converting nebula model to emmeans object
}
