% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_wrappers.R
\name{run_de}
\alias{run_de}
\title{Run DE analysis}
\usage{
run_de(object, pre_de_obj, config_de, ...)
}
\arguments{
\item{object}{object of class giotto or seurat.}

\item{pre_de_obj}{optional object produced by \code{pre_de} or \code{run_pre_de} with cell neighbor expression information.}

\item{config_de}{A list with the following potential configuration parameters:
\itemize{
\item \strong{assay} name of assay to use.
\item \strong{data_use} name of data layer in assay used as outcome in DE.
Could be 'counts' for count based regression or 'normalized' , for a linear based regression.
\item \strong{formula} right hand side of formula used in DE, possibly containing random effects.
Must contain groupVar fixed effect.
It's recommended that raw count models with a log-link (i.e., negative binomial,poisson)
, use a corresponding offset for library size in formula
(i.e., ~de_variable + offset(log(totalcounts))), where totalcounts is a column in the metadata.
\item \strong{groupVar}  DE group variable with two levels,must exist in meta.data slot
\item \strong{groupVar_levels}  optional factor ordering for groupVar, which should have only two-levels (for now)
May be used to control for confounding or 'bleed-over' in DE model (see examples)
\item \strong{family} a family function, a character string naming a family function,
or the result of a call to a family function (variance/link function) information.
See \code{family} for a generic discussion of families.
Default is nbinom2 (with log link), specifying Negative binomial distribution: quadratic parameterization (Hardin & Hilbe 2007). V=mu*(1+mu/phi) = mu+mu^2/phi.
\item \strong{targets} optional vector of targets to run DE for.  If not supplied, run models for all targets in assay.
\item \strong{cellid_colname} column name corresponding to cell id in the metadata of giotto / seurat object.
\item \strong{targets} optional vector of targets to run DE for.  If not supplied, run models for all targets in assay.
\item \strong{ziformula} optional formula for zero inflation model.  Default formula is ~0, corresponding to no zero inflation.
}}

\item{...}{optional further arguments to be passed to during model fitting.
See example below.}
}
\value{
list of lists, pre_de_object can be passed directly to run_de or smi_de functions.
}
\description{
Run DE analysis
}
\examples{
library(Giotto)
datadir<-system.file("extdata", package="smiDE")
gem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
config_pre_de <- list(assay = "rna"
                      ,counts = "raw" 
                      ,normalized = "normalized"
                      ,cellid_colname = "cell_ID"
                      ,cell_type_metadata_colname = "cell_type"
                      ,split_neighbors_by_colname = "tissue"
                      ,mm_radius = 0.05
                      ,ref_celltype = "fibroblast"
                      ,sdimx_colname = "sdimx"
                      ,sdimy_colname = "sdimy"
                      ,weight_colname = "weight"
                      ,contamination = "sum"
                      ,verbose=TRUE
                      )
pre_de_obj <- run_pre_de(gem, config_pre_de) 

config_de <- list(
                  assay = "rna"
                  ,data_use = "raw"
                  ,formula = ~RankNorm(otherct_expr) + niche + tissue + offset(log(totalcounts))
                  ,groupVar="niche"
                  ,ziformula = ~1
                  ,family="nbinom2"
                  ,targets=rownames(gem@expression$rna$raw)[1:2]
                  )

de_obj <- 
  smiDE::run_de(gem, pre_de_obj, config_de)

results(de_obj)

}
