% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_wrappers.R
\name{run_pre_de}
\alias{run_pre_de}
\title{Calculate cell adjacencies and neighboring cell expression of genes before DE analysis.}
\usage{
run_pre_de(object, config_pre_de, ...)
}
\arguments{
\item{object}{object of class giotto or seurat.}

\item{config_pre_de}{A list with the following potential configuration parameters:
\itemize{
\item \strong{assay} name of assay to use.
\item \strong{counts} matrix of counts
\item \strong{normalized} optional, matrix of normalized data.
\item \strong{cellid_colname} cellid_colname column name corresponding to cell id.
\item \strong{cell_type_metadata_colname} column (commonly cell type column) corresponding to categories by which cell expression is aggregated by to calculate expression among neighboring cells by cell_type_metadata_colname type.
\item \strong{split_neighbors_by_colname} (optional) If specified, identification of neighboring cells is split by this column in the meta data.  For example, to avoid identifying cells as neighbors if they were to have close x/y coordinates, but were on different tissue.
\item \strong{mm_radius} maximum euclidean distance in units of sdimx, sdimy by which two cells will be identified as neighbors.
\item \strong{ref_celltype} Reference category (commonly a cell type), used to calculate neighbor expression with respect to.
\item \strong{weight_colname} (optional).  If NULL, neighbor expression is unweighted. If "weight", the column corresponding to "weight" (1/distance) by default is used to weight calculated expression of neighbor cells.
\item \strong{contamination} "sum" by default.
\item \strong{sdimx_colname} column name corresponding to 'x' axis spatial coordinate of cell.
\item \strong{sdimy_colname} column name corresponding to 'y' axis spatial coordinate of cell.
\item \strong{verbose} print some extra messages during calculation.
}}

\item{...}{currently ignored.}
}
\value{
list of lists, pre_de_object can be passed directly to run_de or smi_de functions.
}
\description{
Calculate cell adjacencies and neighboring cell expression of genes before DE analysis.
}
\examples{
library(Giotto)
datadir<-system.file("extdata", package="smiDE")
gem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
config_pre_de <- list(assay = "rna"
                      ,counts = "raw" 
                      ,normalized = "normalized"
                      ,cellid_colname = "cell_ID"
                      ,cell_type_metadata_colname = "cell_type"
                      ,split_neighbors_by_colname = "tissue"
                      ,mm_radius = 0.05
                      ,ref_celltype = "fibroblast"
                      ,sdimx_colname = "sdimx"
                      ,sdimy_colname = "sdimy"
                      ,weight_colname = "weight"
                      ,contamination = "sum"
                      ,verbose=TRUE
                      )
pre_de_obj <- run_pre_de(gem, config_pre_de) 

}
