% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model.R
\name{summarize_model}
\alias{summarize_model}
\title{Function used to summarize models with emmeans based contrasts and a model summary.}
\usage{
summarize_model(
  mod,
  modtime = NULL,
  groupVar = NULL,
  dat = NULL,
  fitmethod,
  original_formula,
  fixed_formula,
  error_msg
)
}
\arguments{
\item{mod}{a model.}

\item{modtime}{optional. An object produced by 'system.time', if provided it is added to the model_summary in returned object.}

\item{groupVar}{optional groupVar.
If specified, this forces pairwise contrasts to be conducted by this variable, even if there are 20 or more levels.}
}
\value{
a list of lists containing data.tables of emmeans summaries.  Intended to be primarily accessed through the smiDE::results function.
}
\description{
Function used to summarize models with emmeans based contrasts and a model summary.
}
