% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Make a volcano plot from smide_results}
\usage{
volcano(
  smide_results,
  comparison = "one.vs.rest",
  variable = NULL,
  targets = "all",
  bonf = NULL,
  lblsize = 4,
  titlval = NULL,
  sig_color_left = "blue",
  sig_color_right = "red",
  fold_change_cutoff = 1.5,
  text_size = 20,
  custom_label_targets = c(),
  interactive = TRUE,
  side_label_bool = FALSE,
  limit_labels = FALSE
)
}
\arguments{
\item{smide_results}{object of class "smide" as returned from \code{smi_de} or \code{run_de} functions.}

\item{comparison}{one of c("pairwise", "one.vs.rest", "one.vs.all").
See ?smiDE::results for details on the individual types of output tables.}

\item{variable}{fixed effect variable in DE model for which to make volcano plot.}

\item{targets}{targets to plot.  typically should be left as the default "all" for full volcano plot.}

\item{bonf}{Optional multiple testing threshold for horizontal dashed line, set by default to 0.05/(# of genes)}

\item{lblsize}{Size of gene labels for significant genes, passed to geom_text_repel.}

\item{titlval}{Optional character string for title.}

\item{sig_color_left}{Color for significant genes with positive negative log fold change.}

\item{sig_color_right}{Color for significant genes with positive log fold change.}

\item{fold_change_cutoff}{fold change cutoff for vertical dashed line in plot indicating a 'large magnitude' of effect, used for indicating whether to color targets as non-grey.}

\item{text_size}{size of plot text set using ggplot2::theme(text=element_text(size=text_size))}

\item{interactive}{TRUE = plotly, FALSE = static ggplot}

\item{side_label_bool}{TRUE = labels nudged to the side of the plot.  FALSE = labels places at datapoint (with lines as determined by ggrepel)}

\item{limit_labels}{FALSE = For non-interactive plots, plot labels for all significant genes with p-value < 0.9*bonf threshold (ggrepel::geom_text_repel max.overlaps argument is set to 1,000).
TRUE = genes with fold_changes with lesser magnitude than \code{fold_change_cutoff} will not be labeled, and ggrepel::geom_text_repel max.overlaps argument will be limited to 100 genes.}
}
\value{
A ggplot2 (volcano) plot object, with log2 fold changes plotted on x-axis, and
}
\description{
Make a volcano plot from smide_results
}
\examples{
library(cowplot)
datadir <- system.file("extdata", package="smiDE")
de_results <- readRDS(paste0(datadir, "/sample_de_results.rds"))

dat <- results(de_results, comparisons="one.vs.rest", variable="group")[["one.vs.rest"]]
head(dat)
volcano(de_results, "one.vs.rest", "group", interactive = TRUE)
volcano(de_results, "pairwise", "group", interactive = FALSE)
plist <- volcano(de_results, "pairwise", "group",text_size=8, lblsize=2, interactive = FALSE)
names(plist)
cowplot::plot_grid(plotlist = plist)

}
